aibot/requirements.txt
# Web framework
Flask

# HTTP requests
requests

# HTML parsing
beautifulsoup4

# Natural Language Processing
textblob

# NLTK (dependency for TextBlob)
NLTK

# Data manipulation
pandas
numpy

# API requests (alternative to requests, included for compatibility)
aiohttp

# Date and time handling
python-dateutil

# Environment variables management
python-dotenv

# WSGI server for production deployment
gunicorn

# Add setuptools to handle the distutils issue
setuptools

# Task scheduling
APScheduler

# Fetching real-time price
yfinance

# MetaApi integration
metaapi-cloud-sdk